Consolidate duplicated tests in the test_rbac_types module and add new tests for list permission type.
