+++ st2common/tests/unit/test_rbac.py
@@ -13,14 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import unittest2
 from oslo_config import cfg
 
 from st2tests import config
 from st2tests.base import CleanDbTestCase
 from st2common.rbac import utils
-from st2common.rbac.types import PermissionType
-from st2common.rbac.types import ResourceType
 from st2common.models.db.auth import UserDB
 from st2common.models.db.rbac import RoleDB
 from st2common.persistence.rbac import Role
@@ -29,7 +26,6 @@ from st2common.rbac.migrations import insert_system_roles
 
 __all__ = [
     'RBACUtilsTestCase',
-    'RBACPermissionTypeTestCase',
     'RBACMigrationsTestCase',
     'RBACRoleDBTestCase'
 ]
@@ -62,41 +58,6 @@ class RBACUtilsTestCase(CleanDbTestCase):
         return
         result = utils.user_is_admin(user_db=self.mocks['user_db'])
         self.assertFalse(result)
-
-
-class RBACPermissionTypeTestCase(unittest2.TestCase):
-    def test_get_valid_permission_for_resource_type(self):
-        valid_action_permissions = PermissionType.get_valid_permissions_for_resource_type(
-            resource_type=ResourceType.ACTION)
-
-        for name in valid_action_permissions:
-            self.assertTrue(name.startswith(ResourceType.ACTION + '_'))
-
-        valid_rule_permissions = PermissionType.get_valid_permissions_for_resource_type(
-            resource_type=ResourceType.RULE)
-
-        for name in valid_rule_permissions:
-            self.assertTrue(name.startswith(ResourceType.RULE + '_'))
-
-    def test_get_resource_type(self):
-        self.assertEqual(PermissionType.get_resource_type(PermissionType.ACTION_CREATE),
-                         ResourceType.ACTION)
-        self.assertEqual(PermissionType.get_resource_type(PermissionType.ACTION_DELETE),
-                         ResourceType.ACTION)
-        self.assertEqual(PermissionType.get_resource_type(PermissionType.ACTION_ALL),
-                         ResourceType.ACTION)
-        self.assertEqual(PermissionType.get_resource_type(PermissionType.PACK_ALL),
-                         ResourceType.PACK)
-
-    def test_get_permission_name(self):
-        self.assertEqual(PermissionType.get_permission_name(PermissionType.ACTION_CREATE),
-                         'create')
-        self.assertEqual(PermissionType.get_permission_name(PermissionType.ACTION_DELETE),
-                         'delete')
-        self.assertEqual(PermissionType.get_permission_name(PermissionType.ACTION_ALL),
-                         'all')
-        self.assertEqual(PermissionType.get_permission_name(PermissionType.PACK_ALL),
-                         'all')
 
 
 class RBACMigrationsTestCase(CleanDbTestCase):
